feat (WebSocket SDK): add cross-platform support; add GIF key and background support; add touch point retrieval support for N4Pro; add custom PID and VID support;
